+++ Sources/AppStoreConnectClient/AppStoreConnectError.swift
@@ -1,3 +1,20 @@
+// FS App Store Connect Client
+// Copyright (C) 2024  FREEDOM SPACE, LLC
+
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published
+//  by the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 //
 //  AppStoreConnectError.swift
 //
@@ -19,8 +36,6 @@ public enum AppStoreConnectError: Error, Equatable {
     case badRequest
     /// The requested resource was not found.
     case notFound
-    /// Invalid access token, indicated by App Store Connect server.
-    case invalidToken
     /// The provided private key is invalid.
     case invalidPrivateKey
     /// The provided payload is invalid.
