+++ Tests/AppStoreConnectClientTests/PayloadTests.swift
@@ -1,3 +1,20 @@
+// FS App Store Connect Client
+// Copyright (C) 2024  FREEDOM SPACE, LLC
+
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published
+//  by the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 //
 //  PayloadTests.swift
 //
@@ -13,7 +30,7 @@ final class PayloadTests: XCTestCase {
     func testPayload() {
         let expiration = Date()
         let expirationClaim = ExpirationClaim(value: expiration)
-        let payload = Payload(
+        let payload = AuthenticationMiddleware.Payload(
             issueID: "issueID",
             expiration: expirationClaim,
             audience: "audience"
